Bump Lua 5.4.7 & sol 3.5.0
Lua is now a static lib and is named "lua.lib".
Sol is updated to current latest which is 3.5.0 at https://github.com/ThePhD/sol2/commit/c1f95a773c6f8f4fde8ca3efe872e7286afe4444

In cargo:
Currently having an issue getting the light color from "Key Light" in `FinalGateEnigma`, making impossible to finish the final enigma.
We're setting the light color of `Key Light` in `OnStart`, we should have it set in `OnAwake` instead. We're trying to access the light color in `FinalGateEnigma` in `OnStart` function, thus leading to Undefined Behavior. The fact it was working previously is pure luck.
Bumping Lua and Sol reavealed this UB due to getting the light color **and** setting the light color at the same time in `OnStart` functions, no ordering was guaranteed.